move around components to satisfy eslint file naming and importing conventions
