+++ src/App.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 import store from './store';
-import Routes from './routes';
-import globalStyles from './app.scss'; // eslint-disable-line no-unused-vars
+import Routes from './Routes';
+import globalStyles from './globalStyles.scss'; // eslint-disable-line no-unused-vars
 
 if (__DEV__) {
   window.store = store;
