+++ src/Routes.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import { Route, Redirect } from 'react-router-dom';
-import Header from './views/header/container';
-import Servers from './views/servers/container';
-import Login from './views/login/container';
+import Header from './views/Header';
+import Servers from './views/Servers';
+import Login from './views/Login';
 
 const PrivateRoutes = connect(state => (
   { token: state.ui.token }
