+++ src/views/Header/index.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import { logout } from '../../actions';
-import Component from './component';
+import Component from './Component';
 
 const mapDispatchToProps = dispatch => ({
   logout: () => dispatch(logout())
