+++ src/views/Servers/index.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import { getServers } from '../../actions';
-import Component from './component';
+import Component from './Component';
 
 const mapStateToProps = state => ({ ...state.servers });
